+++ packages/path-inputs/src/index.ts
@@ -2,3 +2,5 @@ export * from './button';
 export * from './text-input';
 export * from './search-input';
 export * from './hint';
+export * from './typography';
+export * from './tag';