+++ packages/path-kaltura/src/index.ts
@@ -1,0 +1,3 @@
+export * from './kaltura-player-manager';
+export * from './kaltura-player';
+export * from './kaltura-player-context';