+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/kaltura/path-design-system/compare/v0.2.0...v0.3.0) (2020-02-05)
+
+
+### Features
+
+* **path-input:** add tag component  ([#56](https://github.com/kaltura/path-design-system/issues/56)) PATH-486 ([6444d52](https://github.com/kaltura/path-design-system/commit/6444d52a925165c597d8841c37c8a8d0a4014ee8))
+* **path-kaltura:** add kaltura player component Path-455 ([#52](https://github.com/kaltura/path-design-system/issues/52)) ([4d353e6](https://github.com/kaltura/path-design-system/commit/4d353e6ed5ea2fa4d2dc0f8e0c94501e2594fdd8))
+
+
+
+
+
 # [0.2.0](https://github.com/kaltura/path-design-system/compare/v0.1.2...v0.2.0) (2020-01-16)